Avoid out-of-bounds array accesses
The VREG file implementation used a ternary operator to select array
elements based on the value of a genvar. Some simulators and/or
synthesis tools considered this to be an out-of-bounds access, even if
the out-of-bounds value is not actually used. Replace this instance by
an if/else block. Fixes #8.

Additionally, out-of-bounds array accesses occured when implementing
the registers with Xilinx's RAM32 primitive if the number of internal
read ports was not a multiple of 3. Add additional conditional blocks
to avoid these out-of-bounds accesses as well.

No functional changes.