Move busy flag out of state structs
The busy flag, which is also used to indicate whether the other members
of a state struct are valid, needs to be reset, while the other members
do not. Move the busy flag out of the state structs in order to avoid
toolchain issues when attempting to reset just one member of a struct.
Potential fix for #6. No functional changes.