Fix VREG write port multiplexing delay count
Fix calculation of the VREG write port multiplexing delay count.  Some
units must issue write commands to the vector register file for several
cycles due to multiplexing of the write ports.  The maximum write delay
was calculated incorrectly, causing overly pessimistic delays which
required additional buffers and hence a waste of resources.  Also, add
assertions to check that specified maximum number of write attempts is
compatible with the throughput of the respective units.