Add ready flags to stages of ALU, MUL, SLD, and ELEM
This is the first patch to support stall logic in the ALU, MUL, SLD,
and ELEM units.  Ready flags are added, with data moving from one stage
to the next only if the target stage is ready.  However, currently the
last stage is always ready, which propagates through the pipeline and
means that all ready flags are always set.