Add new pending read/writes interface to units
Add a new interface for pending vector register (vreg) reads and writes
to units.  This consists of a local copy of the global write mask and a
new output port with pending reads for each unit.

The local copy of the global write mask is updated with the current
global mask for every new instruction entering a unit and subsequently
a logical AND with the global mask is performed each cycle.  This has
the effect that vregs are successively cleared in the local copy but
new vregs of later instructions added to the global mask are not added
to the local copy.  That way the local copy only contains vregs written
to by previous instructions and can be used to test for data hazards
when reading.

The new output port with pending reads informs other units about the
vregs that will be read by a unit.  It does not contain any of the
vregs set in the local write mask copy of that unit, hence all vregs in
that output mask are "owned" by that unit at that time and will
eventually be read by that unit.  Other units must stall before writing
to any vreg contained in the pending read mask of any unit.

This commit only adds the functionality of the interface but that
functionality is not used yet.