Use XIF memory interface
Replace the current memory interface with the XIF memory interface.
Currently, the top module does not wire the memory interface through
the main core but keeps using a dedicated memory interface for the
vector coprocessor.  Hence, this commit does not introduce functional
changes but merely replaces the existing interface with the
standardized XIF memory interface.