Fix illegal vreg logic in decoder for reductions
Reduction instructions read the initial value from a single vreg and
write their result to a single vreg, rather than a vreg group.  Prevent
the decoder from flagging these as invalid vreg addresses for LMUL > 1.
Modify some of the reduction tests to verify this (by using vregs that
are not aligned to the LMUL setting as initial and result vregs).