Add EVL policy to decoder
Add a policy for determining the effective vector length (EVL) of
vector instructions.  The default policy is to use the VL set by the
previous `vset[i]vl[i]` instruction, scaled by the EEW/SEW ratio for
vector loads and stores.  However, some instructions require setting
the EVL to 1, to the maximum supported by the current config, or to
VL/8 for vector mask loads and stores.