Combine ALU mask flag and op mask field
ALU operations cannot be masked of the mask vreg is used as part of the
operation (i.e., as carry in or as a selector).  Hence, the mask flag
can be combined with the `op_mask` field of the ALU mode.

In addition, the ALU mode is extended with a `inv_res` flag that allows
to invert the result, similar to how the `inv_op*` fields allow the
inversion of operands.  This can be used to influence the effect of the
carry in subtraction operations by inverting either one of the operands
and then inverting the result if needed.