Replace ALU `VAADD` with `shift_op` flag & introduce VSADD
Use the VADD operation of the vector ALU for averaging add and subtract
instructions by shifting the operands and generating carry in based on
the rounding mode in case a new `shift_op` flag is asserted.  This
replaces the obsolete `VAADD` operation.  Instead, a dedicated `VSADD`
operation is added, which is used by saturating add and subtract
instructions.