Introduce common pipeline and move ALU to it
This is the first patch in a series that will move most of the redundant
pipeline code in the individual units into a common pipeline module.
It starts by adding a new pipeline module and stripping the ALU of most
of the pipeline code that is moved to that new module.