Adapt unit counter width to smallest operand
The granularity of the unit counter should be chosen such that it allows
counting the smallest operand of a unit.  In particular, if the unit
supports element-wise operation, then the smallest operand width is
always 8 bits.  As a consequence, the stride counter part of the LSU is
merged with the regular low part.