Add alt counter and remove obsolete debug signals
Add definition for an alternative counter flag that allows to indicate
that certain operands use an alternative counter, with the intent of
using that for the SLD operand.  Initialize the alternative counter for
use with the SLD unit and increment it along with the regular counter.
Also, remove some obsolete debug signals.