Add preliminary unit mux
Add a preliminary implementation for a unit multiplexer, which allows
several units to be placed in one pipeline.  This preliminary version
adds the base functionality required for routing pipeline data to one
of several units and to forward results from several units.  However,
support is currently still limited to a single unit per pipeline.