Fix corner case of operand fetch logic
Fetching operands is inhibited if a unit requires the auxiliary counter
and the value of that counter is not currently 0.  However, the flag
indicating whether the auxiliary counter is used could be outdated in
case the pipeline is currently accepting a new instruction.  Yet, in
the first cycle of any instruction the auxiliary counter is either
unused or 0, and thus fetching should take place.