Prepare pipeline flushing logic in unit wrapper
Some instructions executed by the ELEM unit require flushing of the
downstream part of the pipeline to flush write data into the target
vector register.  So far, the entire pipeline executed a dummy
instruction for this purpose.  Prepare the required logic to move the
flushing into the unit wrapper in the future.