Use a dynamic SV package for default config values
This commit introduces some major changes to the way the vector
coprocessor is configured.  Instead of separately parsing and feeding
configuration values via top-level parameters to the design for each
synthesis or simulation workflow, a new centralized Makefile is used to
dynamically generate a SV package with consistent default configuration
values that are the default parameter values of the vproc_core module.
These values can still be overriden when instantiating the core module.