Fix unit mux attempting to stall the VLSU
The VLSU cannot be stalled (since the memory response interface cannot
be stalled either).  However, with other units sharing the VLSU's
pipeline the unit mux could attempt to stall the LSU if another unit
provides its results first.  To fix this, the unit mux now keeps track
of the number of instructions currently within the LSU and stalls other
units until the LSU is done with all of them.