Switch to LLVM as default compiler
Make LLVM the default compiler (and search for clang >= 14 in $PATH) in
the `sw/` Makefile.  Do not set the `COMPILER` env variable in the
`test/` Makefile (it is not required there and will be set in the `sw/`
Makefile anyways.  Switch to Ubuntu 22.04's LLVM 14 package for CI
(avoids compiling a suitable RISC-V compiler from source).