Ensure that instr stay in order within a pipeline
In a pipeline some units may be faster than others, which could
potentially change the order of instructions within a unit, creating
data dependency issues (e.g., an instruction waiting for data from
another instruction while simultaneously denying that other instruction
access to the output pipe of the unit mux).  Introduce a queue in the
unit mux to ensure that units get access to the output pipe in the same
order they received their instructions.