Do not abort Verilator simulation by default
Avoid aborting Verilator simulation after there was no activity on the
memory interface by default.  Some programs may not access the memory
for long periods of time.  Make aborting the simulation optional
instead, controlled by the environment variable `SIM_ABORT_CYCLES`,
that may be used to specify the number of cycles of inactivity after
which simulation is aborted.  Issue a warning if simulation is aborted.