Add timing-predictability to config.mk
Add a new environment variable `TIMEPRED` to the configuration Makefile
to control whether Vicuna shall be configured in a timing-predictable
fashion.  As a first measure, absence of this env variable relaxes the
vector register write port multiplexing policy and allows stalling
pipelines in case of vector register write port collisions for
non-timing-predictable configurations.